dissolve routeHooks
