+++ extension/src/panel/zodiac-routes/index.ts
@@ -1,4 +1,7 @@
-export * from './routeHooks'
 export { useSelectedRouteId } from './SelectedRouteContext'
 export { useZodiacRoute } from './useZodiacRoute'
-export { ProvideZodiacRoutes, useZodiacRoutes } from './ZodiacRouteContext'
+export {
+  ProvideZodiacRoutes,
+  useMarkRouteAsUsed,
+  useZodiacRoutes,
+} from './ZodiacRouteContext'
